build: Do not build libtraceevent always
It had PHONY dependency which means that it will be built always.
Although Makefile in the directly will handle the dependency, it'd be
better to check it here so that it can skip the recursive make.

It's actually useful for upcoming build change to use (another)
recursive make call to handle dependency correctly.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>