build: Suppress warning from (parallel) make
If configure is not called before running make, it will internally run
configure to check dependencies.  But if the make was called with -j
option, following warning is shown.

  make[1]: warning: jobserver unavailable: using -j1.  Add '+' to parent make rule.

I think it's because the configure runs (recursive) make in it and the
child make will see MAKEFLAGS (env?) variable has -j option.  We don't
want parallel build here, so just remove the flags.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>